Guarantee your own commits via a blocking payer (#441)
The goal of this PR is to not return a commit (or any blockchain
payload) as successfull to the client until at least one node
acknowledged the write. This is important for clients as they like to
re-read their writes.

This functionality uses the new metadata API to check when the correct
sequenceID reached the desired state.