Allow running without payer key (#547)
If you were to try to run the XMTP API/SYNC without providing the payer
key, the node will crash:
```
$ ./dev/run
2025-02-24T14:01:43.485-0500    INFO    replication     Version: v0.2.0-9-g1f8192f
2025-02-24T14:01:43.495-0500    INFO    replication     Successfully connected to DB    {"namespace": "xmtpd_c04b2306614f"}
2025-02-24T14:01:43.512-0500    FATAL   replication     initializing signer     {"error": "unable to parse private key: invalid length, need 256 bits"}
exit status 1
```

The blockchain publisher is only required for the payer. If we ever need
it for the node, it will have to use the node key.

I will break apart the XMTPD image and the payer image in the next few
days. But I need this to get the infra code moving.