Verifiable payer (#544)
Uses https://github.com/xmtp/proto/pull/247

This adds a new field to the payer envelope that prevents a node from
stealing payloads https://github.com/xmtp/xmtpd/issues/474

Additionally it gets rid of old AAD validation #463

The bulk of the changeset allows the payer to pick a different node if
the connection to the ideal node has failed. This is obviously not
ideal, because it will keep retrying even if the node is down long term,
but in the short/medium term this is sufficient.

Fixes #474 #463