Print registrant in expected form (#613)
This makes reading datadog logs much easier. The utils function dumps
the string in the same format as our CLI and what we get out of the
contracts.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Enhanced system logging to ensure essential operational events are
captured more effectively for improved monitoring.

In this release we've updated how key system events are recorded. This
improvement aids in more reliable tracking and diagnostics without
affecting your overall experience or visible functionality.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->